+++ Assets/Scripts/UIHealth.cs
@@ -9,12 +9,12 @@ public class HealthBar : MonoBehaviour
     public Gradient gradient;
     public Image fill;
     // Start is called before the first frame update
-    public void SetMaxHealth(int health){
+    public void SetMaxHealth(float health){
         slider.maxValue = health;
         slider.value = health;
         fill.color = gradient.Evaluate(1f);
     }
-    public void SetHealth(int health){
+    public void SetHealth(float health){
         slider.value = health;
         fill.color = gradient.Evaluate(slider.normalizedValue);
     }
